Move core of LowerLayersPass into separate class
Rather than have state in the Pass object itself, which needs to be cleared
after running, move state out into an object which is constructed fresh for
each run of the pass. This way, we don't have pointers to stack-allocated
locals in the pass, or any state that we might forget to clear.